Add user Location & Permissions
The user Location was added delegating the logic
to a separate viewModel, in which we ask/handle permissions
in multiple cases and use a integrated delegate which runs
everytime a locationManager is created/instantiated to tell when
the permissions have changed. Also updated the Plist to allow
location while in app use.