Add user location centering & some refactoring
The initial region was refactored to be in the ViewModel
so that the ui changes when change is detected. Since showUserLocation is true
the map will have that location as soon as the permission is given. so on the
switch case the region is re set to the maps region and thus the user location
is shown. Moved the ViewModel to own file for better readability, created a
default enum to DRY up the code.